a step towards refactoring into submodules
this commit is BROKEN.
