+++ markdown2canvas/free_functions/__init__.py
@@ -1,0 +1,597 @@
+import os.path as path
+import os
+
+import logging
+logger = logging.getLogger(__name__)
+
+import canvasapi
+
+def is_file_already_uploaded(filename,course):
+    """
+    returns a boolean, true if there's a file of `filename` already in `course`.
+
+    This function wants the full path to the file.
+    """
+    return ( not find_file_in_course(filename,course) is None )
+
+
+
+
+def find_file_in_course(filename,course):
+    """
+    Checks to see of the file at `filename` is already in the "files" part of `course`.
+
+    It tests filename and size as reported on disk.  If it finds a match, then it's up.
+
+    This function wants the full path to the file.
+    """
+    import os
+
+    base = path.split(filename)[1]
+
+    files = course.get_files()
+    for f in files:
+        if f.filename==base and f.size == path.getsize(filename):
+            return f
+
+    return None
+
+
+
+
+
+def is_page_already_uploaded(name,course):
+    """
+    returns a boolean indicating whether a page of the given `name` is already in the `course`.
+    """
+    return ( not find_page_in_course(name,course) is None )
+
+
+def find_page_in_course(name,course):
+    """
+    Checks to see if there's already a page named `name` as part of `course`.
+
+    tests merely based on the name.  assumes assignments are uniquely named.
+    """
+    import os
+    pages = course.get_pages()
+    for p in pages:
+        if p.title == name:
+            return p
+
+    return None
+
+
+
+def is_assignment_already_uploaded(name,course):
+    """
+    returns a boolean indicating whether an assignment of the given `name` is already in the `course`.
+    """
+    return ( not find_assignment_in_course(name,course) is None )
+
+
+def find_assignment_in_course(name,course):
+    """
+    Checks to see if there's already an assignment named `name` as part of `course`.
+
+    tests merely based on the name.  assumes assingments are uniquely named.
+    """
+    import os
+    assignments = course.get_assignments()
+    for a in assignments:
+
+        if a.name == name:
+            return a
+
+    return None
+
+
+
+
+def get_canvas_key_url():
+    """
+    reads a file using an environment variable, namely the file specified in `CANVAS_CREDENTIAL_FILE`.
+
+    We need the
+
+    * API_KEY
+    * API_URL
+
+    variables from that file.
+    """
+    from os import environ
+
+    cred_loc = environ.get('CANVAS_CREDENTIAL_FILE')
+    if cred_loc is None:
+        raise SetupError('`get_canvas_key_url()` needs an environment variable `CANVAS_CREDENTIAL_FILE`, containing the full path of the file containing your Canvas API_KEY, *including the file name*')
+
+    # yes, this is scary.  it was also low-hanging fruit, and doing it another way was going to be too much work
+    with open(path.join(cred_loc),encoding='utf-8') as cred_file:
+        exec(cred_file.read(),locals())
+
+    if isinstance(locals()['API_KEY'], str):
+        logger.info(f'using canvas with API_KEY as defined in {cred_loc}')
+    else:
+        raise SetupError(f'failing to use canvas.  Make sure that file {cred_loc} contains a line of code defining a string variable `API_KEY="keyhere"`')
+
+    return locals()['API_KEY'],locals()['API_URL']
+
+
+def make_canvas_api_obj(url=None):
+    """
+    - reads the key from a python file, path to which must be in environment variable CANVAS_CREDENTIAL_FILE.
+    - optionally, pass in a url to use, in case you don't want the default one you put in your CANVAS_CREDENTIAL_FILE.
+    """
+
+    key, default_url = get_canvas_key_url()
+
+    if not url:
+        url = default_url
+
+    return canvasapi.Canvas(url, key)
+
+
+
+def generate_course_link(type,name,all_of_type,courseid=None):
+    '''
+    Given a type (assignment or page) and the name of said object, generate a link
+    within course to that object.
+    '''
+    if type in ['page','quiz']:
+        the_item = next( (p for p in all_of_type if p.title == name) , None)
+    elif type == 'assignment':
+        the_item = next( (a for a in all_of_type if a.name == name) , None)
+    elif type == 'file':
+        the_item = next( (a for a in all_of_type if a.display_name == name) , None)
+        if the_item is None: # Separate case to allow change of filenames on Canvas to names that did exist
+            the_item = next( (a for a in all_of_type if a.filename == name) , None)
+            # Canvas retains the name of the file uploaded and calls it `filename`. 
+            # To access the name of the document seen in the Course Files, we use `display_name`.
+    else:
+        the_item = None
+
+        
+    if the_item is None:
+        print(f"ℹ️ No content of type `{type}` named `{name}` exists in this Canvas course.  Either you have the name incorrect, the content is not yet uploaded, or you used incorrect type before the colon")
+    elif type == 'file' and not courseid is None:
+        # Construct the url with reference to the coruse its coming from
+        file_id = the_item.id
+        full_url = the_item.url
+        stopper = full_url.find("files")
+
+        html_url = full_url[:stopper] + "courses/" + str(courseid) + "/files/" + str(file_id)
+
+        return html_url
+    elif type == 'file':
+        # Construct the url - removing the "download" portion
+        full_url = the_item.url
+        stopper = full_url.find("download")
+        return full_url[:stopper]
+    else:
+        return the_item.html_url
+    
+
+
+def find_in_containing_directory_path(target):
+    import pathlib
+
+    target = pathlib.Path(target)
+
+    here = pathlib.Path('.').absolute()
+
+    testme = here / target
+
+    found = testme.exists()
+
+    while (not found) and here.parent!=here:
+        here = here.parent
+        testme = here / target
+        found = testme.exists()
+
+
+    if not found:
+        raise FileNotFoundError('unable to find {} in a containing folder of {}'.format(target, pathlib.Path('.').absolute()))
+
+    return here / target
+
+
+
+def preprocess_replacements(contents, replacements_path):
+    """
+    attempts to read in a file containing substitutions to make, and then makes those substitutions
+    """
+
+    if replacements_path is None:
+        return contents
+    with open(replacements_path,'r',encoding='utf-8') as f:
+        import json
+        replacements = json.loads(f.read())
+
+    for source, target in replacements.items():
+        contents = contents.replace(source, target)
+
+    return contents
+
+
+    
+
+def preprocess_markdown_images(contents,style_path):
+
+    rel_style_path = find_in_containing_directory_path(style_path)
+
+    contents = contents.replace('$PATHTOMD2CANVASSTYLEFILE',str(rel_style_path))
+
+    return contents
+
+
+def get_default_property(key, helpstr):
+
+    defaults_name = find_in_containing_directory_path(path.join("_course_metadata","defaults.json"))
+
+    try:
+        logger.info(f'trying to use defaults from {defaults_name}')
+        with open(defaults_name,'r',encoding='utf-8') as f:
+            import json
+            defaults = json.loads(f.read())
+
+        if key in defaults:
+            return defaults[key]
+        else:
+            print(f'no default `{key}` specified in {defaults_name}.  add an entry with key `{key}`, being {helpstr}')
+            return None
+
+    except Exception as e:
+        print(f'WARNING: failed to load defaults from `{defaults_name}`.  either you are not at the correct location to be doing this, or you need to create a json file at {defaults_name}.')
+        return None
+
+
+def get_default_style_name():
+    return get_default_property(key='style', helpstr='a path to a file relative to the top course folder')
+
+def get_default_replacements_name():
+    return get_default_property(key='replacements', helpstr='a path to a json file containing key:value pairs of text-to-replace.  this path should be expressed relative to the top course folder')
+
+
+
+
+def apply_style_markdown(sourcename, style_path, outname):
+    from os.path import join
+
+    # need to add header and footer.  assume they're called `header.md` and `footer.md`.  we're just going to concatenate them and dump to file.
+
+    with open(sourcename,'r',encoding='utf-8') as f:
+        body = f.read()
+
+    with open(join(style_path,'header.md'),'r',encoding='utf-8') as f:
+        header = f.read()
+
+    with open(join(style_path,'footer.md'),'r',encoding='utf-8') as f:
+        footer = f.read()
+
+
+    contents = f'{header}\n{body}\n{footer}'
+    contents = preprocess_markdown_images(contents, style_path)
+
+    with open(outname,'w',encoding='utf-8') as f:
+        f.write(contents)
+
+
+
+
+def apply_style_html(translated_html_without_hf, style_path, outname):
+    from os.path import join
+
+    # need to add header and footer.  assume they're called `header.html` and `footer.html`.  we're just going to concatenate them and dump to file.
+
+    with open(join(style_path,'header.html'),'r',encoding='utf-8') as f:
+        header = f.read()
+
+    with open(join(style_path,'footer.html'),'r',encoding='utf-8') as f:
+        footer = f.read()
+
+
+    return f'{header}\n{translated_html_without_hf}\n{footer}'
+
+
+
+
+
+def markdown2html(filename, course, replacements_path):
+    """
+    This is the main routine in the library.
+
+    This function returns a string of html code.  
+
+    It does replacements, emojizes, converts markdown-->html via `markdown.markdown`, and does page, assignment, and file reference link adjustments.
+
+    If `course` is None, then you won't get some of the functionality.  In particular, you won't get link replacements for references to other content on Canvas.
+
+    If `replacements_path` is None, then no replacements, duh.  Otherwise it should be a string or Path object to an existing json file containing key-value pairs of strings to replace with other strings.
+    """
+    if course is None:
+        courseid = None
+    else:
+        courseid = course.id
+
+    root = path.split(filename)[0]
+
+    import emoji
+    import markdown
+    from bs4 import BeautifulSoup
+
+
+    with open(filename,'r',encoding='utf-8') as file:
+        markdown_source = file.read()
+
+    markdown_source = preprocess_replacements(markdown_source, replacements_path) 
+
+    emojified = emoji.emojize(markdown_source)
+
+
+    html = markdown.markdown(emojified, extensions=['codehilite','fenced_code','md_in_html','tables','nl2br']) # see https://python-markdown.github.io/extensions/
+    soup = BeautifulSoup(html,features="lxml")
+
+    all_imgs = soup.findAll("img")
+    for img in all_imgs:
+            src = img["src"]
+            if ('http://' not in src) and ('https://' not in src):
+                img["src"] = path.join(root,src)
+
+    all_links = soup.findAll("a")   
+    course_page_and_assignments = {}
+    if any(l['href'].startswith("page:") for l in all_links) and course:
+        course_page_and_assignments['page'] = course.get_pages()
+    if any(l['href'].startswith("assignment:") for l in all_links) and course:
+        course_page_and_assignments['assignment'] = course.get_assignments()
+    if any(l['href'].startswith("quiz:") for l in all_links) and course:
+        course_page_and_assignments['quiz'] = course.get_quizzes()
+    if any(l['href'].startswith("file:") for l in all_links) and course:
+        course_page_and_assignments['file'] = course.get_files()
+    for f in all_links:
+            href = f["href"]
+            root_href = path.join(root,href)
+            split_at_colon = href.split(":",1)
+            if path.exists(path.abspath(root_href)):
+                f["href"] = root_href
+            elif course and split_at_colon[0] in ['assignment','page','quiz','file']:
+                type = split_at_colon[0]
+                name = split_at_colon[1].strip()
+                get_link = generate_course_link(type,name,course_page_and_assignments[type],courseid)
+                if get_link:
+                    f["href"] = get_link
+                    
+
+    return str(soup)
+
+
+
+
+
+def find_local_images(html):
+    """
+    constructs a map of local url's : Images
+    """
+    from bs4 import BeautifulSoup
+
+    soup = BeautifulSoup(html,features="lxml")
+
+    local_images = {}
+
+    all_imgs = soup.findAll("img")
+
+    if all_imgs:
+        for img in all_imgs:
+            src = img["src"]
+            if src[:7] not in ['https:/','http://']:
+                local_images[src] = Image(path.abspath(src))
+
+    return local_images
+
+
+
+
+
+def adjust_html_for_images(html, published_images, courseid):
+    """
+    
+    published_images: a dict of Image objects, which should have been published (so we have their canvas objects stored into them)
+
+    this function edits the html source, replacing local url's
+    with url's to images on Canvas.
+    """
+    from bs4 import BeautifulSoup
+
+    soup = BeautifulSoup(html,features="lxml")
+
+    all_imgs = soup.findAll("img")
+    if all_imgs:
+        for img in all_imgs:
+            src = img["src"]
+            if src[:7] not in ['https:/','http://']:
+                # find the image in the list of published images, replace url, do more stuff.
+                local_img = published_images[src]
+                img['src'] = local_img.make_src_url(courseid)
+                img['class'] = "instructure_file_link inline_disabled"
+                img['data-api-endpoint'] = local_img.make_api_endpoint_url(courseid)
+                img['data-api-returntype'] = 'File'
+    
+    return str(soup)
+
+    # <p>
+    #     <img class="instructure_file_link inline_disabled" src="https://uws-td.instructure.com/courses/3099/files/219835/preview" alt="hauser_menagerie.jpg" data-api-endpoint="https://uws-td.instructure.com/api/v1/courses/3099/files/219835" data-api-returntype="File" />
+    # </p>
+
+
+
+
+def find_local_files(html):
+    """
+    constructs a list of BareFiles, so that they can later be replaced with a url to a canvas thing
+    """
+    from bs4 import BeautifulSoup
+
+    soup = BeautifulSoup(html,features="lxml")
+
+    local_files = {}
+
+    all_links = soup.findAll("a")
+
+    if all_links:
+        for file in all_links:
+            href = file["href"]
+            if path.exists(path.abspath(href)):
+                local_files[href] = BareFile(path.abspath(href))
+
+    return local_files
+
+
+
+def adjust_html_for_files(html, published_files, courseid):
+
+
+    # need to write a url like this :
+    # <a class="instructure_file_link instructure_scribd_file" title="airport.csv" href="https://uws.instructure.com/courses/501155/files/47304001/download?wrap=1" data-api-endpoint="https://uws.instructure.com/api/v1/courses/501155/files/47304001" data-api-returntype="File">Download</a>
+
+
+    from bs4 import BeautifulSoup
+
+    soup = BeautifulSoup(html,features="lxml")
+
+    all_files = soup.findAll("a")
+
+    if all_files:
+        for file in all_files:
+            href = file["href"]
+            if path.exists(path.abspath(href)):
+                # find the image in the list of published images, replace url, do more stuff.
+                local_file = published_files[href]
+                file['href'] = local_file.make_href_url(courseid)
+                file['class'] = "instructure_file_link instructure_scribd_file"
+                file['title'] = local_file.name # what it's called when you download it???
+                file['data-api-endpoint'] = local_file.make_api_endpoint_url(courseid)
+                file['data-api-returntype'] = 'File'
+
+    return str(soup)
+
+
+
+def get_root_folder(course):
+    for f in course.get_folders():
+        if f.full_name == 'course files':
+            return f
+
+
+
+
+
+
+
+
+
+def get_assignment_group_id(assignment_group_name, course, create_if_necessary=False):
+
+    existing_groups = course.get_assignment_groups()
+
+    if not isinstance(assignment_group_name,str):
+        raise RuntimeError(f'assignment_group_name must be a string, but I got {assignment_group_name} of type {type(assignment_group_name)}')
+
+
+    for g in existing_groups:
+        if g.name == assignment_group_name:
+            return g.id
+
+
+
+    if create_if_necessary:
+        msg = f'making new assignment group `{assignment_group_name}`'
+        logger.info(msg)
+
+        group = course.create_assignment_group(name=assignment_group_name)
+        group.edit(name=assignment_group_name) # this feels stupid.  didn't i just request its name be this?
+
+        return group.id
+    else:
+        raise DoesntExist(f'cannot get assignment group id because an assignment group of name {assignment_group_name} does not already exist, and `create_if_necessary` is set to False')
+
+    
+
+
+
+def create_or_get_assignment(name, course, even_if_exists = False):
+
+    if is_assignment_already_uploaded(name,course):
+        if even_if_exists:
+            return find_assignment_in_course(name,course)
+        else:
+            raise AlreadyExists(f"assignment {name} already exists")
+    else:
+        # make new assignment of name in course.
+        return course.create_assignment(assignment={'name':name})
+
+
+
+def create_or_get_page(name, course, even_if_exists):
+    if is_page_already_uploaded(name,course):
+
+        if even_if_exists:
+            return find_page_in_course(name,course)
+        else:
+            raise AlreadyExists(f"page {name} already exists")
+    else:
+        # make new assignment of name in course.
+        result = course.create_page(wiki_page={'body':"empty page",'title':name})
+        return result
+
+
+
+
+def create_or_get_module(module_name, course):
+
+    try:
+        return get_module(module_name, course)
+    except DoesntExist as e:
+        return course.create_module(module={'name':module_name})
+
+
+
+
+def get_module(module_name, course):
+    """
+    returns 
+    * Module if such a module exists, 
+    * raises if not
+    """
+    modules = course.get_modules()
+
+    for m in modules:
+        if m.name == module_name:
+            return m
+
+    raise DoesntExist(f"tried to get module {module_name}, but it doesn't exist in the course")
+
+
+def get_subfolder_named(folder, subfolder_name):
+
+    assert '/' not in subfolder_name, "this is likely broken if subfolder has a / in its name, / gets converted to something else by Canvas.  don't use / in subfolder names, that's not allowed"
+
+    current_subfolders = folder.get_folders()
+    for f in current_subfolders:
+        if f.name == subfolder_name:
+            return f
+
+    raise DoesntExist(f'a subfolder of {folder.name} named {subfolder_name} does not currently exist')
+
+    
+def delete_module(module_name, course, even_if_exists):
+
+    if even_if_exists:
+        try:
+            m = get_module(module_name, course)
+            m.delete()
+        except DoesntExist as e:
+            return
+
+    else:
+        # this path is expected to raise if the module doesn't exist
+        m = get_module(module_name, course)
+        m.delete()
